requirements.txt
asgiref==3.7.2
Django==3.2.20
django-rest-passwordreset==1.3.0
djangorestframework==3.12.4
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
